ecksum: Adopted global constant in lfs3_gbmap_set_ when ecksum is NULL
It's two words, and I've probably already spent too long fiddling around
with this.

This drops the messy NULL checks in lfs3_gbmap_set_, for a wrapper that
defaults to a global {.cksize=-1} ecksum when NULL.

---

Code changes, with both a compound literal (cl), and global
constant (gc). The current code uses a global constant:

                      code          stack          ctx
  before:            35144           2136          660
  after+cl:          35144 (+0.0%)   2136 (+0.0%)  660 (+0.0%)
  after+gc:          35144 (+0.0%)   2136 (+0.0%)  660 (+0.0%)

                      code          stack          ctx
  gbmap+np before:   38272           2144          776
  gbmap+np after+cl: 38412 (+0.4%)   2144 (+0.0%)  776 (+0.0%)
  gbmap+np after+gc: 38380 (+0.3%)   2144 (+0.0%)  776 (+0.0%)

                      code          stack          ctx
  gbmap+yp before:   38940           2168          796
  gbmap+yp after+cl: 38952 (+0.0%)   2168 (+0.0%)  796 (+0.0%)
  gbmap+yp after+gc: 38920 (-0.1%)   2168 (+0.0%)  796 (+0.0%)

It's interesting to note that while the global constant generally
reduces code cost, it prevents constant-expr optimizations from
eliminating the NULL checks when compiling without preerases (np).

Is that enough reason to revert this? Probably not. (1) The simpler
codebase, and reduced chance of forgetting a NULL check, is preferable,
and (2) we don't care about the code cost of niche gbmap configurations
as much as the non-gbmap modes.