Fixed weakened wear-leveling (mdir aliasing) due to rev changes
This is a common pitfall with mdirs that has only ended up in the
codebase ~3(?) times. Naively using a power-of-two recycle counter for
relocations ends up aliasing mdir blocks such that only one block is
actually wear-leveled.

I'm not entirely sure it was intentional, but the previous
double-increment during needsrelocation checks made mdirs relocate one
recycle early, avoiding this aliasing issue.

However, the double-increment had other issues. The most glaring is that
it would always trigger two relocations back-to-back due to the mismatch
between counter cycles and overflow checks. Sort of defeating the
purpose of wear-leveling the two mdir blocks separately...

---

What we really want is a counter that's always coprime with 2. Such as
our old friend mod (2^n)-1.

Unfortunately mod (2^n)-1 counters don't really have any great
optimization trick. They show up all the time when code relies on the
multiplicative cycle of a 2^n finite-field, but despite this, all of
the implementations I've seen rely on a simple branch to handle the
one extra state.

I'm not sure this is the best implementation, but adding 2 and
subtracting 1 on non-overflow seems to minimize resulting code cost.
Presumably because the compiler is able to deduplicate this with the
needsrelocation check. Though we're only talking about a handful of
bytes.

Code changes:

                    code          stack          ctx
  before:          35152           2136          660
  after:           35164 (+0.0%)   2136 (+0.0%)  660 (+0.0%)

                    code          stack          ctx
  gbmap before:    38388           2144          776
  gbmap after:     38400 (+0.0%)   2144 (+0.0%)  776 (+0.0%)

                    code          stack          ctx
  preerase before: 38928           2168          796
  preerase after:  38940 (+0.0%)   2168 (+0.0%)  796 (+0.0%)

---

We don't have great tests for this, as it's difficult to create
rigorous checks for our best-effort dynamic wear-leveling. But,
surprisingly enough, this _was_ caught by
test_exhaustion_spam_uzd_fuzz's doubling-disk-doubles-lifetime check!

Though only with LFS3_YES_GBMAP=1:

  LFS3_YES_GBMAP=1 \
          TESTS=tests/test_exhaustion.toml \
          make test-runner -j \
      && ./scripts/test.py test_exhaustion_spam_uzd_fuzz -O- -j \
      | grep lifetime